Delete callback-based APIs for events, properties, and replies.
They are poorly designed and not terribly useful. I wrote the original
versions of these libraries for demonstration purposes and would like to
actively discourage anyone from actually using them.

After deleting the callback-based APIs, there was nothing interesting
left in property or reply, so those libraries are deleted outright. The
event library is no longer particularly related to event handling, but
that's a problem for another commit.

The icccm library had some simple hooks for integrating with the
property library, which are precisely as useful as the property library
itself, so I deleted them too. (Arnaud suggested this in
<sa5ocib84hf.fsf@Orfeo.duckcorp.org>.)

Since the icccm and event libraries have changed interfaces, this commit
bumps their SONAME versions.

Signed-off-by: Jamey Sharp <jamey@minilop.net>
Reviewed-by: Mikhail Gusarov <dottedmag@dottedmag.net>
Reviewed-by: Josh Triplett <josh@joshtriplett.org>
Acked-by: Julien Cristau <jcristau@debian.org>